+++ crates/rec_aggregation/tests/test_log2_ceil.py
@@ -1,5 +1,5 @@
 from snark_lib import *
-from ..utils import *
+from ..zkdsl_implem.utils import *
 
 
 def main():
